requirements.txt
Django==2.2.13
django-widget-tweaks==1.4.3

django-phone-field==1.8.1
django-bootstrap-modal-forms==2.2.0

# For image fields and form generator
Pillow==8.2.0

langdetect==1.0.9
openpyxl==3.0.7

# Django global login required middleware
django-glrm==1.1.3

# For date convert
jdatetime==3.6.2

# For Excel Export
xlsxwriter==1.1.2


# Cron through admin interface, BEST EVER!! but incompatible with django 2.x
django-chroniker==1.0.17

# Deployment PostgreSQL
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1

jdatetime==3.6.2

# For Wizards
# Was part of Django till Django-1.7
django-formtools==2.2

# Celery
celery==4.4.0
redis==3.3.11

# date time localization
babel==2.8.0
